feat: improve array and object types with default values
